1.33.0.0_CT: Added #1942

+++ source/res/scripts/client/MusicControllerWWISE.py
@@ -123,8 +123,7 @@ class MusicController(object):
         def replace(self, event, eventId, unlink=True):
             if self.__event is not None:
                 if self.__event.name != event.name:
-                    if unlink:
-                        self.__event.stop()
+                    if self.__eventID != MUSIC_EVENT_LOBBY and unlink:
                         self.__event.unlink()
                 else:
                     self.__eventID = eventId
